created navbar component and side navbar component for users view
