+++ client/src/components/index.js
@@ -6,7 +6,7 @@ export { default as CaseDetails } from './CaseDetails';
 export { default as DeactivateAccountButton } from './DeactivateAccountButton';
 export { default as MediatorList } from './MediatorList';
 export { default as NavBar } from './NavBar';
-export { default as SettingsNavBar } from './SettingsNavBar';
+export { default as SideNavBlock } from './SideNavBlock';
 export { default as UserCaseList } from './UserCaseList';
 export { default as MediatorPendingCaseList } from './MediatorPendingCaseList';
 export { default as MediatorAcceptedCaseList } from './MediatorAcceptedCaseList';
