+++ client/src/styles/index.scss
@@ -27,6 +27,7 @@
 @import './navbar.scss';
 @import './home.scss';
 @import './mediator-cases-show.scss';
+@import './side-nav-block.scss';
 @import './grid/grid.scss';
 @import './grid/flex.scss';
 @import './grid/spacing.scss';
