+++ client/src/styles/navbar.scss
@@ -6,3 +6,10 @@
     z-index:10;
     position: absolute;
 }
+
+.braveHeader {
+    display: block;
+    color: white;
+   text-align: center;
+   margin: 0;
+}
