+++ client/src/styles/side-nav-block.scss
@@ -1,0 +1,8 @@
+.lowerNavBlock {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color:$bg-brav-primary;
+}
