+++ client/src/styles/variables.scss
@@ -76,7 +76,8 @@ $margin-sizes: (
 /**
  * Colors
  */
-
+ 
+$bg-brav-primary: #5C90C1;
 $color-default: black;
 $color-success: green;
 $color-secondary: grey;
